refactor: decompose cadis-core/lib.rs by extracting helper modules
